feat: write bean to firestore
closes #85
